feat(analytics): enhance visitor analytics with EarthGlobe component and new assets
- Introduced the EarthGlobe component for visualizing visitor data on a 3D globe, utilizing @react-three/fiber and @react-three/drei for rendering.
- Added new assets: earth-blue-marble.jpg, earth-night.jpg, and earth-topology.png for the globe's texture.
- Updated VisitorGlobe component to integrate EarthGlobe and improve visitor data representation.
- Enhanced visitor popover with detailed information including location, last seen time, and device details.

This commit significantly improves the analytics interface by providing an interactive 3D visualization of visitor data.

+++ crates/temps-analytics/src/analytics.rs
@@ -2859,7 +2859,7 @@ WHERE project_id = $1
                 COUNT(*) FILTER (WHERE e.is_entry = true) as entry_count,
                 COUNT(*) FILTER (WHERE e.is_exit = true) as exit_count,
                 COUNT(*) FILTER (WHERE e.is_bounce = true) as bounce_count,
-                AVG(e.time_on_page) FILTER (WHERE e.time_on_page IS NOT NULL AND e.time_on_page > 0) as avg_time_on_page
+                (AVG(e.time_on_page) FILTER (WHERE e.time_on_page IS NOT NULL AND e.time_on_page > 0))::float8 as avg_time_on_page
             FROM events e
             WHERE {}
             GROUP BY e.page_path
